Fix Gyro hangs
The I2C communication in the Wire.h library is implemented with while loops. If the physical wiring of the I2C component introduces stale readings from the pins the functions could stick in the while loop so the function never returns. This would cause the hang of the whole FW of the OAT.

Changes:
* Implemented the timeout feature in the FW so it cannot hang anymore
* Set the clock rate of the I2C communication to the lowest setting to increase stability with imperfect wiring (too long cables, no pull-up resistors, etc.)
* If a timeout happens it is now logged on the INFO level
* Fixed some typos